Add a registerNotifySignals to directly connect notify signals to lambdas
The current way to emit dataChanged is to connect signals' QMetaMethod
to onItemPropertyChanged however when invoked this function has to call
QObject::sender() and QObject::senderSignalIndex to figure which object
and property changed. Under the hood this locks a mutex (twice) and has
proven to be a performance bottleneck.
By connecting directly to a lambda the object and property associated
with a signal can be captured thus avoiding the sender and
senderSignalIndex calls.
This is unfortunately not transparent for the library user and has to be
explicitly called.